requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
